LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-data-url-sub-frame.html from frame with URL data:text/html,<html><head><script>function test() {try {parent.document.getElementById('accessMe').innerHTML = 'FAIL: Cross frame access from a data: URL was allowed.';} catch (e) {}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body onload="test()"><p>Inner iframe.</p></body></html>. Domains, protocols and ports must match.

This tests that a data: URL loaded in an iframe doesn't have access to its parent's frame


PASS: Cross frame access from a data: URL was denied.

--------
Frame: 'aFrame'
--------
Inner iframe.
